% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{aa_counts}
\alias{aa_counts}
\title{Compute Alcohol Attributable Counts}
\usage{
aa_counts(aaf_table, dh)
}
\arguments{
\item{aaf_table}{a tibble of AAFs as returned by compute_aafs.}

\item{dh}{a formatted deaths/hosps table as returned by format_v*_dh}
}
\description{
Applies the computed AAFs to user provided death/hospitalization counts.
}
