% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{aaf_calibration_factory}
\alias{aaf_calibration_factory}
\title{Calibrated AAF Factory}
\usage{
aaf_calibration_factory(IM, COUNT, DRINKERS, N_GAMMA, LB, BB, UB)
}
\arguments{
\item{IM}{InterMAHP condition code}

\item{COUNT}{Count of condition events over the relevant time period}

\item{DRINKERS}{Estimated # of drinkers in population}

\item{N_GAMMA}{Normalized gamma distribution (used as exposure mass function)}

\item{LB}{lower, binge, and upper bounds of consumption}

\item{BB}{lower, binge, and upper bounds of consumption}

\item{UB}{lower, binge, and upper bounds of consumption}
}
\value{
AAF_CMP function for distributing harm for wholly attributable
functions
}
\description{
Invokes cond_prob_factory and wraps the result multiplied by th
}
