% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{aaf_cmp_factory}
\alias{aaf_cmp_factory}
\title{Factory for alcohol attributable fraction computing functions}
\usage{
aaf_cmp_factory(LB, UB, RR_FD, P_FD, INTGRND)
}
\arguments{
\item{LB}{dbl, lower bound of consumption}

\item{UB}{dbl, upper bound of consumption}

\item{RR_FD}{dbl, relative risk for former drinkers}

\item{P_FD}{dbl, proportion of former drinkers}

\item{INTGRND}{fn, continuous function to be integrated}
}
\value{
a function object that takes as input a vector x of values >= 0.03 and
 returns the alcohol attributable fraction between 0.03 and x for outcome
 OUTCOME due to condition CONDITION among cohort GENDER * AGE_GROUP in the
 place and time REGION * YEAR
}
\description{
Given function data (gamma distribution specs, relative risk
 curves, binge ratios) aaf_factory produces a function f that computes aaf
 aafs from a given interval
}
