% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02_aaf_factory.R
\name{aaf_fd}
\alias{aaf_fd}
\title{Computes alcohol attributable fraction for former drinkers}
\usage{
aaf_fd(aaf_specs)
}
\arguments{
\item{aaf_specs}{is a tibble that contains the variables:
LB: dbl
UB: dbl
P_FD: dbl
RR_FD: dbl
INTGRND: fn}
}
\value{
double to be interpreted as a fraction
}
\description{
Given risk, population, and current drinker integrand information
 aaf_fd computes the alcohol attributable fraction among former drinkers
}
