% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02-postprocessors.R
\name{aaf_total}
\alias{aaf_total}
\title{Compute AAFs for all conditions for current drinkers and totals AAFs}
\usage{
aaf_total(aaf_table)
}
\arguments{
\item{aaf_table}{A tibble as returned by assemble (i.e. has the following vars
AAF_CMP: fn, VIVO, computes AAF from LB to each input
AAF_FD: dbl, AAF for former drinkers
UB: dbl, consumption upper bound
)}
}
\description{
Compute AAFs for all conditions for current drinkers and totals AAFs
}
