% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaf_evaluation.R
\name{add_cutpoints}
\alias{add_cutpoints}
\title{Add evaluation cutpoints to a given datatable}
\usage{
add_cutpoints(aaf_table, cuts)
}
\arguments{
\item{aaf_table}{a tibble as returned by assemble}

\item{cuts}{a list of double vectors indexed by aaf_table$GENDER}
}
\description{
Written to provide additional reactivity to Shiny InterMAHP app.
}
