% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{addFormerFraction}
\alias{addFormerFraction}
\title{Computes a given scenario's AAF for former drinkers and adds it to the
scenario}
\usage{
addFormerFraction(.data, var_name = "aaf_fd")
}
\description{
Computes a given scenario's AAF for former drinkers and adds it to the
scenario
}
