% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{addGenderStratifiedIntervalFraction}
\alias{addGenderStratifiedIntervalFraction}
\title{Compute a given scenario's AAF for current drinkers in the given intervals of
consumption, stratified over the given values of the gender variable, and add
it to the scenario.}
\usage{
addGenderStratifiedIntervalFraction(.data, lower_strata, upper_strata,
  var_name = "aaf_xd")
}
\arguments{
\item{.data}{a scenario}

\item{var_name}{the name of the new variable to be added}

\item{strata}{a list where names(list) intersects with .data$gender, and each
entry of strata is a list with a "lower" and "upper" bound of consumption}
}
\description{
Compute a given scenario's AAF for current drinkers in the given intervals of
consumption, stratified over the given values of the gender variable, and add
it to the scenario.
}
