% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{addIntervalFraction}
\alias{addIntervalFraction}
\title{Compute a given scenario's AAF over the given interval (not gender
 stratified) and add it to the scenario}
\usage{
addIntervalFraction(.data, lower, upper, var_name = "aaf_xd")
}
\description{
Compute a given scenario's AAF over the given interval (not gender
 stratified) and add it to the scenario
}
