% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{addIntervalFractions}
\alias{addIntervalFractions}
\title{Compute a given scenario's AAF over the given intervals (not gender
 stratified) and add it to the scenario}
\usage{
addIntervalFractions(.data, lower, upper, grp_names)
}
\description{
Compute a given scenario's AAF over the given intervals (not gender
 stratified) and add it to the scenario
}
