% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{assemble}
\alias{assemble}
\title{Collect and assemble AAF data from RR and PC input data}
\usage{
assemble(pc, rr, ext, lb, ub, bb, gc = list(Female = 1.258^2, Male = 1.171^2))
}
\arguments{
\item{pc}{Prevalence / Consumption input}

\item{rr}{Relative Risk input}

\item{ext}{logical, extrapolate linearly?}

\item{lb}{Double, consumption lower bound}

\item{ub}{Double, consumption upper bound}

\item{bb}{Double vector, Binge consumption level, Gender stratified}
}
\description{
Collect and assemble AAF data from RR and PC input data
}
