% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{binge_risk_factory}
\alias{binge_risk_factory}
\title{Produce a Relative-Risk-For-Bingers curve from the given input}
\usage{
binge_risk_factory(IM, BINGEF, LNXT_RR)
}
\arguments{
\item{IM}{string, Intermahp condition code}

\item{BINGEF}{double, rescales injuries}

\item{LNXT_RR}{continuous vector valued function as returned by lin_ext_fact}
}
\value{
a function whose values are binge modified.  This affects conditions
 5.2 and 5.5, ischaemic heart disease and stroke respectively, by removing a
 protective effect at low levels of consumption
}
\description{
given a list that contains the string variable IM and the double
 variable BINGEF, produces a Relative-Risk-For-Bingers curve.  IM is used for
 the curves for IHD and ischaemic stroke, where any protective J-shape is
 forfeited by bingers.  All curves are then multiplied by BINGEF which has
 the value of 1 in all cases except for condition classes 7, 8, 9, which
 account for motor vehicle collisions, unintentional injuries, and
 intentional injuries.
}
