% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{bng_rr}
\alias{bng_rr}
\title{Produce a Relative-Risk-For-Bingers curve from the given input}
\usage{
bng_rr(rr_specs, fn_rr)
}
\arguments{
\item{rr_specs}{a list that contains the string variables IM and the double
BINGEF}

\item{rr_fn}{a relative risk curve}
}
\description{
given a list that contains the string variable IM and the double
 variable BINGEF, produces a Relative-Risk-For-Bingers curve.  IM is used for
 the curves for IHD and ischaemic stroke, where any protective J-shape is
 forfeited by bingers.  All curves are then multiplied by BINGEF which has
 the value of 1 in all cases except for condition classes 7, 8, 9, which
 account for motor vehicle collisions, unintentional injuries, and
 intentional injuries.
}
