% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{calibrate}
\alias{calibrate}
\title{Combine Prev/Cons and Death/Hosp data to calibrate wholly attributable curves}
\usage{
calibrate(dh)
}
\arguments{
\item{dh}{as returned by derive_v*_dh}
}
\description{
Wholly attributable conditions have an AAF_TOTAL of 1.00, but we can still
distribute this mass over the interval of interest (i.e. 0.03 to UB).
When applicable (4.(1, 2, 3), 6.(1, 5) we calibrate a loglinear conditional
probability curve whose area is equal to the regional yearly prevalence of
the given condition among current drinkers.
}
