% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{calibration_factory}
\alias{calibration_factory}
\title{Produce a Conditional Probability curve from the given input}
\usage{
calibration_factory(dh_specs)
}
\arguments{
\item{rr_specs}{is a tibble that contains the variables:
IM: chr
COUNT: dbl
DRINKERS: dbl
N_GAMMA: fn
LB: dbl
BB: dbl
UB: dbl}
}
\value{
a function whose values are binge modified.  This affects conditions
 5.2 and 5.5, ischaemic heart disease and stroke respectively, by removing a
 protective effect at low levels of consumption
}
\description{
Given a list that contains the true count among a given cohort of
a wholly alocohol-attributable condition, the number of drinkers in that
cohort, the conditions's IM code, and the gamma function detailing consumption
among the cohort, this function calibrates a slope for a loglinear function
that estimates the conditional probablity of developing the given condition.
The function should be interpreted as conditional probability mass.  I.e., the
integral of the function over the relevant range is the probability that a
drinker will be afflicted by an event of the given condition over the period
of time that the given Count variable was collected.

Uses a nonlinear optimizer (COBYLA) to find a loglinear slope for the
function f(x) = max(1, exp(k(x-t))) that mimizes the difference between
integral(N_GAMMA * (f-1), 0.03, UB) and yearly prevalence (count/drinkers).

The goal is to produce a continuous analogue to the relative risk curve for
conditions that are wholly attributable to alcohol. The assumption is made
that such a condition has a loglinear thresholded (i.e. f(x)=1 for x<t)
conditional probablity function on the interval of concern (0.03 to UB grams
of ethanol/day, averaged over 1yr).

This conditional probability is used to portion a 1.00 AAF_TOTAL among the
drinking population.
}
