% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkAttributability}
\alias{checkAttributability}
\title{Check the attributability of a given vector of intermahp codes}
\usage{
checkAttributability(im)
}
\description{
Uses internal master attributability list (see rr_master)
}
