% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkVars}
\alias{checkVars}
\title{Checks variable names}
\usage{
checkVars(.data, expected)
}
\arguments{
\item{.data}{a data_frame object}

\item{expected}{character vector of expected variables}
}
\description{
Checks against the given list, errors if a variable is missing and returns
only the specified variables.
}
