% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean}
\alias{clean}
\title{Clean data}
\usage{
clean(.data, expected)
}
\arguments{
\item{.data}{a data_frame object}

\item{expected}{character vector of expected variables}
}
\value{
The cleaned dataset
}
\description{
Performs universal cleaning operations of lowering variable name case,
ensuring necessary variables, imputing missing values when possible, and
reconverting types after imputation.
}
