% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{compile_rr}
\alias{compile_rr}
\title{Compile a list of RR functions and the parameters that produced them}
\usage{
compile_rr(R)
}
\arguments{
\item{R}{a list that contains the string variables IM, GENDER, and FUNCTION,
sixteen double variables B1 through B16, the logical EXT, and the double
BINGEF.}
}
\description{
Compile a list of RR functions and the parameters that produced them
}
