% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aafs.R
\name{compute_aaf}
\alias{compute_aaf}
\title{Combine curve and prevalence/consumption data into an aaf}
\usage{
compute_aaf(C, P)
}
\arguments{
\item{C}{is an entry in the list curve of the form:
[[1]]: a list [IM, Condition, Gender, Outcome]
[[2]]: Relative risk for former drinkers, double
[[3]]: Relative risk for current nonbinging drinkers, function
[[4]]: Relative risk for current binge drinkers, function}

\item{P}{is a row from the data.table params with entries:
  Region  Year    Gender  Age_Group Population
  PCC_g/d Cor.Fac Rel.Con P_LA

We don't need those first nine to define integrands.  The remaining are:
  P_FD        P_CD    P_BD    drinker PCC.drk
  Binge bndry G.shape G.scale nc      df
  p_bat       R1      R2}
}
\description{
Combine curve and prevalence/consumption data into an aaf
}
