% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaf_evaluation.R
\name{compute_aafs}
\alias{compute_aafs}
\title{Compute AAFs for all conditions as separated by the given cutpoints}
\usage{
compute_aafs(aaf_table, cuts)
}
\arguments{
\item{aaf_table}{A tibble as returned by assemble}

\item{cuts}{A list of double vectors indexed by aaf_table$GENDER}
}
\description{
Compute AAFs for all conditions as separated by the given cutpoints
}
