% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{compute_all_aafs}
\alias{compute_all_aafs}
\title{Compute All AAFs}
\usage{
compute_all_aafs(prev_cons_data = intermahpr::pc_default,
  relative_risks = intermahpr::rr_default, extrapolation = TRUE,
  upper_bound = 250, binge_levels = c(53.8, 67.25),
  light_moderate_barriers = c(20, 20), moderate_heavy_barriers = c(40, 40))
}
\arguments{
\item{PC}{Prevalence and consumption}

\item{RR}{Relative risk informer}
}
\value{
data.frame of aafs by region, year, gender, age group, condition
}
\description{
Compute all AAFs.
}
\examples{
compute_all_aafs(intermahpr::pc_default, intermahpr::rr_default, TRUE, c())

}
