% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{compute_general_aafs}
\alias{compute_general_aafs}
\title{Compute General AAFs}
\usage{
compute_general_aafs(prev_cons_data = intermahpr::pc_default,
  relative_risks = intermahpr::rr_default, extrapolation = TRUE,
  upper_bound = 250, binge_levels = c(53.8, 67.25), lower = c(0.03, 0.03),
  upper = c(upper_bound, upper_bound))
}
\arguments{
\item{PC}{Prevalence and consumption}

\item{RR}{Relative risk informer}
}
\value{
data.frame of aafs by region, year, gender, age group, condition
}
\description{
Interface functions are wrappers for this more general aaf compiler
}
\examples{
 compute_general_aafs(intermahpr::pc_default,
                    intermahpr::rr_default,
                    TRUE,
                    250,
                    c(53.80, 67.25))


}
