% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{compute_heavy_aafs}
\alias{compute_heavy_aafs}
\title{Compute Heavy AAFs}
\usage{
compute_heavy_aafs(prev_cons_data = intermahpr::pc_default,
  relative_risks = intermahpr::rr_default, extrapolation = TRUE,
  upper_bound = 250, binge_levels = c(53.8, 67.25),
  moderate_heavy_barriers = c(35, 35))
}
\arguments{
\item{PC}{Prevalence and consumption}

\item{RR}{Relative risk informer}
}
\value{
data.frame of aafs by region, year, gender, age group, condition
}
\description{
Compute heavy aafs.
}
\examples{
compute_heavy_aafs(intermahpr::pc_default, intermahpr::rr_default, TRUE, c())

}
