% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aafs.R
\name{compute_interval_aaf_deprecated}
\alias{compute_interval_aaf_deprecated}
\title{Compute AAF numerators over a given interval from a given RR curve / Probabil
ity Dist.}
\usage{
compute_interval_aaf_deprecated(R1, R2, PR, PB, bb, lb, ub)
}
\arguments{
\item{R1}{The ratio (P_CD - P_BD) / (P_CD - P_BAT)}

\item{R2}{The ratio (P_BD - P_BAT)/ (P_CD - P_BAT)}

\item{PR}{The function P(x)(RR(X)-1), P is adjusted gamma distribution, RR is
relative risk}

\item{PB}{The function P(X)(RB(X)-1), P is adjusted gamma distribution, RB is
binge-adjusted relative risk}

\item{bb}{The binge drinking definition (g/day)}

\item{lb}{Lower bound of interval}

\item{ub}{Upper bound of interval}
}
\description{
This method computes, in full generality, the numerator terms found in
Section 4.1 of the InterMAHP Comprehensive guide.  Currently unused, as that
integral fits on a single line.
}
