% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_total_aafs.R
\name{compute_total_aafs}
\alias{compute_total_aafs}
\title{Compute Total AAFs}
\usage{
compute_total_aafs(PC = get_prevalence_consumption_data(),
  RR = get_relative_risk_data())
}
\arguments{
\item{PC}{Prevalence and consumption}

\item{RR}{Relative risk informer}
}
\value{
data.frame of aafs by region, year, gender, age group, condition
}
\description{
Compute them.
}
\examples{
 compute_total_aafs(intermahpr::pc_default, intermahpr::rr_default)

}
