% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{computeIntervalFraction}
\alias{computeIntervalFraction}
\title{Compute a given scenario's AAF for current drinkers in a given interval of
consumption}
\usage{
computeIntervalFraction(.data, lower = -Inf, upper = Inf)
}
\description{
Compute a given scenario's AAF for current drinkers in a given interval of
consumption
}
