% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02-factories.R
\name{cond_prob_factory}
\alias{cond_prob_factory}
\title{Conditional Probability Factory}
\usage{
cond_prob_factory(IM, COUNT, DRINKERS, N_GAMMA, LB, BB, UB)
}
\arguments{
\item{IM}{InterMAHP condition code}

\item{COUNT}{Count of condition events over the relevant time period}

\item{DRINKERS}{Estimated # of drinkers in population}

\item{N_GAMMA}{Normalized gamma distribution (used as exposure mass function)}

\item{LB}{lower, binge, and upper bounds of consumption}

\item{BB}{lower, binge, and upper bounds of consumption}

\item{UB}{lower, binge, and upper bounds of consumption}
}
\value{
Conditional probability mass function for risk incurred as a result of
exposure
}
\description{
Invokes slope_calibration and wraps the result in a loglinear
probability mass function
}
