% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{copy_aaf_cmp_factory}
\alias{copy_aaf_cmp_factory}
\title{Produces a copy of a given AAF computing function}
\usage{
copy_aaf_cmp_factory(aaf_specs)
}
\arguments{
\item{aaf_specs}{is a tibble that contains the variables:
AAF_CMP: fn}
}
\description{
This is a function because the given AAF computing function apparently needs
an environment buffer.  Also normalizes the code in join_dh_aaf
}
