% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rr.R
\name{crushBetas}
\alias{crushBetas}
\title{Crush numbered variables}
\usage{
crushBetas(.data)
}
\description{
All numbered variables (i.e. b1-b16) are crushed into a single variable.
Assumes numbered variables are al of numeric type.
New variable is named 'betas'.
}
