% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{deduce_relative_risk_curves_from_rr}
\alias{deduce_relative_risk_curves_from_rr}
\title{Get a list of Former Drinker Relative Risks, Relative Risk Curves, and Binge
Risk Curves}
\usage{
deduce_relative_risk_curves_from_rr(RR = intermahpr::rr_default,
  x_in = TRUE)
}
\arguments{
\item{RR}{A data table of relative risk information formatted as demanded in
the InterMAHP comprehensive guide}

\item{extrapolation}{Either TRUE(linear) or FALSE(capped) used for
extrapolating the RR after 150 (100 for IHD)}
}
\description{
Get a list of Former Drinker Relative Risks, Relative Risk Curves, and Binge
Risk Curves
}
