% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{derive_dh}
\alias{derive_dh}
\title{Derives count data for use in calibration and portioning}
\usage{
derive_dh(dh, pc)
}
\arguments{
\item{dh}{Deaths/Hosps as returned by format_dh}

\item{pc}{Prev/Cons as returned by derive_pc}
}
\description{
Derives count data for use in calibration and portioning
}
