% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{derive_v0_dh}
\alias{derive_v0_dh}
\title{Derives count data for use in calibration and portioning}
\usage{
derive_v0_dh(dh, pc)
}
\arguments{
\item{dh}{Deaths/Hosps as returned by format_v0_dh}

\item{pc}{Prev/Cons as returned by derive_v0_pc}
}
\description{
-Melts Deaths/Hosps variables into outcome/count (provides better abstraction
for calibration methods)
-Derives as count/drinkers the region/year/cohort prevalence among drinkers
-
}
