% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{derive_v0_rr}
\alias{derive_v0_rr}
\title{Derives relative risk curves from relative risk data}
\usage{
derive_v0_rr(rr, ext)
}
\arguments{
\item{rr}{tibble as formatted above}

\item{ext}{logical indicator of extrapolation, (TRUE = linear, FALSE = capped)}
}
\value{
tibble which is RR with an additional column that contains all
 relevent relative risk curves under the variable CURVES. The CURVES variable
 has the following structure:
   RR[["CURVES"]][[i]] is a list
   RR[["CURVES"]][[i]][["BASE_RR"]] is the base relative risk function (i.e.
     no extrapolation after 100/150)
   RR[["CURVES"]][[i]][["LNXT_RR"]] is the extrapolated relative risk curve
   RR[["CURVES"]][[i]][["BNGD_RR"]] is the extrapolated relative risk curve
     for binge drinkers
}
\description{
Derives relative risk curves from relative risk data
}
