% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{derive_v1_pc}
\alias{derive_v1_pc}
\title{Derives probability distribution parameters from Prevalence and Consumption
data}
\usage{
derive_v1_pc(.data, bb = list(Female = 53.8, Male = 67.25), lb = 0.03,
  ub = 250, gc = list(Female = 1.258^2, Male = 1.171^2))
}
\arguments{
\item{.data}{tibble as formatted in format_v*_pc}

\item{bb}{Gender stratified binge barrier}

\item{lb}{Lower bound}

\item{ub}{Upper bound}

\item{gc}{Gender stratified gamma constants

Constants used:
 0.002739726 = 1/365
 0.7893 = unit conversion (millilitres to grams) for ethanol}
}
\value{
A tibble with additional variables:
  PCC_G_DAY: Per capita consumption, converted from litres/year to grams/day
  DRINKERS: Total drinkers in region, computed as POPULATION*P_CD
  PCAD: Average consumption among drinkers
  PCC_AMONG_DRINKERS: Computed from PCAD and relative consumption
  GAMMA_CONSTANT: The distribution of drinkers at drinking level of x grams
    per day is fit by a gamma distribution whose mean and standard deviation
    are related by GAMMA_CONSTANT, stratified by gender
  GAMMA_SHAPE: Computed from GAMMA_CONSTANT
  GAMMA_SCALE: Computed from GAMMA_CONSTANT
  BB: Binge barrier
  LB: Lower bound of integration (0.03)
  UB: Upper bound of integration (user-specified, typical values are 150,250)
  NC: Normalizing constant for the gamma distribution (how much of the gamma
  distribution is between LB and UB?)
  DF: Deflation factor for the gamma distribution (DF = P_CD/NC)
  P_BAT: Proportion of bingers above threshold (BB)
  R1, R2: Ratios computed at this step and used in AAF computations
}
\description{
From a given prevalence and consumption dataframe (or similar), computes re-
-quired data for AAF computation including Per Capita Consumption among
drinkers, and Gamma Shape/Scale parameters.  Also computes constants
necessary for AAF integration, namely the Deflation factor and Binge-Split
ratios.
}
\details{
Gamma parameters are derivable entirely from prevalence and consumption,
no additional input is necessary.
}
