% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dh.R
\docType{data}
\name{dh_vars}
\alias{dh_vars}
\title{List of variables expected to be in a DH sheet}
\format{An object of class \code{character} of length 7.}
\usage{
dh_vars
}
\description{
List of variables expected to be in a DH sheet
}
\keyword{datasets}
