% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{distillModel}
\alias{distillModel}
\title{Extract and derive pertinent comparative scenario data from a model}
\usage{
distillModel(.data)
}
\description{
Extract and derive pertinent comparative scenario data from a model
}
