% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_aaf_evaluation.R
\name{evaluate_at_cutpoints}
\alias{evaluate_at_cutpoints}
\title{Evaluate a given datatable at pre-added cutpoints}
\usage{
evaluate_at_cutpoints(aaf_table_cuts)
}
\arguments{
\item{aaf_table_cuts}{a tibble as returned by add_cutpoints}
}
\description{
Written to provide additional reactivity to shiny app
}
