% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{ext_rr}
\alias{ext_rr}
\title{Produce an extrapolation-modified relative risk curve}
\usage{
ext_rr(rr_specs)
}
\arguments{
\item{rr_specs}{is a tibble that contains the variables:
IM: chr
GENDER: chr
EXT: lgl}
}
\value{
a function whose values above 150 (100 for IHD) are linearly
 extrapolated (see InterMAHP guide for details)
}
\description{
given a relative risk curve and some metadata, produces a
 risk curve that has the desired extrapolation behaviour beyond 150 g/day,
 or 100 g/day for IHD.
}
