% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{extract_prevcons}
\alias{extract_prevcons}
\title{Helper function that tidies base AAF into the relevant prevalence and
consumption data}
\usage{
extract_prevcons(aaf_table)
}
\arguments{
\item{aaf_table}{a tibble as returned by add_cutpoints with cutpoints of the
form c(LB, LM, MH, UB) where LM is the light-moderate barrier and MH is the
moderate-heavy barrier.}
}
\description{
Intended for use with intermahp base functionality (assumes the positions of
cutpoint variables).  Exported for use in the Shiny app, called behind the
scenes when invoking intermahp_base.
}
