% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{FemaleAcutePancreatitisSpline}
\alias{FemaleAcutePancreatitisSpline}
\title{The spline for Female Acute Pancreatitis}
\usage{
FemaleAcutePancreatitisSpline(x)
}
\description{
0.04 = 1/25
0.0007304602 = 1/(37^2)
}
