% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{fit_slope}
\alias{fit_slope}
\title{Slope Solver for Thresholded Loglinear Functions}
\usage{
fit_slope(pc_dh_specs)
}
\arguments{
\item{pc_dh_specs}{is a tibble that contains the variables:
IM: chr
COUNT: dbl
N_GAMMA: fn
LB: dbl
BB: dbl
UB: dbl
DRINKERS: dbl}
}
\description{
Uses a nonlinear optimizer (COBYLA) to find a loglinear slope for the
function f(x) = max(1, exp(k(x-t))) that mimizes the difference between
integral(N_GAMMA * (f-1), 0.03, UB) and yearly prevalence (count/drinkers).
}
\details{
The goal is to produce a continuous analogue to the relative risk curve for
conditions that are wholly attributable to alcohol. The assumption is made
that such a condition has a loglinear thresholded (i.e. f(x)=1 for x<t)
conditional probablity function on the interval of concern (0.03 to UB grams
of ethanol/day, averaged over 1yr).

This conditional probability is used to portion a 1.00 AAF_TOTAL among the
drinking population.
}
