% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{format_pc}
\alias{format_pc}
\title{Format prevalence and consumption input data to our desired specifications}
\usage{
format_pc(.data)
}
\arguments{
\item{.data}{is a dataset that can be converted to tibble with the expected
variables}
}
\description{
Formats all variable names to upper case and imputes imputable missing data.
Intended for a prevalence and consumption sheet as described in the intermahp
user guide.

expected variables:
   P_FD (Proportion, Former Drinkers)
   P_BD (Proportion, Binge Drinkers)
   P_LA (Proportion, Lifetime Abstainers)
   P_CD (Proportion, Current Drinkers)
   Gamma Distribution Parameters:
     PCC_litres_year (Per Capita Consumption)
     Population
     Relative_Consumption
     Correction_Factor
   Gender (in order to pair with RR curves)
   Region
   Year
   Gender
   Age_Group
}
