% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{format_v0_dh}
\alias{format_v0_dh}
\title{Format Count Data}
\usage{
format_v0_dh(dh)
}
\arguments{
\item{dh}{is a dataset that can be converted to tibble with the expected
variables}
}
\description{
Formats all variable names to upper case and imputes imputable missing data.
Intended for a count sheet

expected variables:
   IM (Used to join this data with aafs)
   Region
   Year
   Gender
   Age_Group
   Outcome
   Count
}
