% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{format_v0_rr}
\alias{format_v0_rr}
\title{Format relative risk input data to our desired specifications}
\usage{
format_v0_rr(rr)
}
\description{
Format relative risk input data to our desired specifications
}
