% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{format_v1_rr}
\alias{format_v1_rr}
\title{Format relative risk input data to our desired specifications}
\usage{
format_v1_rr(.data)
}
\arguments{
\item{.data}{is a dataset that can be converted to tibble with the expected
variables}
}
\description{
Formats all variable names to upper case and imputes imputable missing data.
Intended for a relative risk sheet as described in the intermahp user guide.

expected variables:
 Gender (paired with PC, DH data)
 Function (FP, Spline, with valid IM, or Step, for HIV only)
 B1-B16 (only if Function == FP. need nonzero betas, rest can be missing)
 IM
 Condition
 Outcome
 RR_FD
 BingeF
}
