% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{fp_rr}
\alias{fp_rr}
\title{Get Fractional Polynomial Relative Risk Function}
\usage{
fp_rr(betas)
}
\arguments{
\item{B}{The numeric vector of Beta values needed to produce a fractional
polynomial}
}
\description{
Get Fractional Polynomial Relative Risk Function
}
