% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{fractional_polynomial_rr}
\alias{fractional_polynomial_rr}
\title{Get Fractional Polynomial Relative Risk Function}
\usage{
fractional_polynomial_rr(betas)
}
\arguments{
\item{betas}{The numeric vector of Beta values needed to produce a fractional
polynomial}
}
\description{
Get Fractional Polynomial Relative Risk Function
}
