% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\docType{data}
\name{fractional_polynomials}
\alias{fractional_polynomials}
\title{The master list of fractional polynomials}
\format{An object of class \code{list} of length 16.}
\usage{
fractional_polynomials
}
\description{
The master list of fractional polynomials
}
\keyword{datasets}
