% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{gamma_factory}
\alias{gamma_factory}
\title{Factory for gamma distributions}
\usage{
gamma_factory(GAMMA_SHAPE, GAMMA_SCALE)
}
\arguments{
\item{GAMMA_SHAPE, GAMMA_SCALE}{Gamma parameters to be supplied to dgamma}
}
\description{
We want easier access to gamma distributions with a wide array of fixed params
}
