% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermahpr_constants.R
\name{get_female_light_moderate_barrier}
\alias{get_female_light_moderate_barrier}
\title{Get and Set drinking group bounds}
\usage{
get_female_light_moderate_barrier()
}
\description{
Get and Set drinking group bounds
}
