% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermahpr_constants.R
\name{get_relative_risk_data}
\alias{get_relative_risk_data}
\title{Get and Set relative risk and prevalence/consumption data}
\usage{
get_relative_risk_data()
}
\description{
Get and Set relative risk and prevalence/consumption data
}
