% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermahpr_constants.R
\name{get_upper_bound}
\alias{get_upper_bound}
\title{Get and Set upper and lower bounds of integration}
\usage{
get_upper_bound()
}
\description{
Get and Set upper and lower bounds of integration
}
