% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getExpectedVars_}
\alias{getExpectedVars_}
\title{Get variables expected to be in the given object type}
\usage{
getExpectedVars_(.obj_type)
}
\description{
Get variables expected to be in the given object type
}
