% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02-factories.R
\name{\%prod\%}
\alias{\%prod\%}
\title{Pointwise Function Product Factory as Binary Operator}
\usage{
f \%prod\% g
}
\arguments{
\item{f}{function that takes a single argument and produces a value that is
a valid argument for the `*` function}

\item{g}{function that takes a single argument and produces a value that is
a valid argument for the `*` function}
}
\description{
binary operator for product_factory
}
