% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{hiv_f_rr}
\alias{hiv_f_rr}
\title{Special Female HIV Relative Risk Function}
\usage{
hiv_f_rr(x)
}
\arguments{
\item{x}{A vector of x values at which to evaluate the RR function}
}
\description{
Special Female HIV Relative Risk Function
}
