% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{hiv_m_rr}
\alias{hiv_m_rr}
\title{Special Male HIV Relative Risk Function}
\usage{
hiv_m_rr(x)
}
\arguments{
\item{x}{A vector of x values at which to evaluate the RR function}
}
\description{
Special Male HIV Relative Risk Function
}
