% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{hypertension_m_rr}
\alias{hypertension_m_rr}
\title{Special Spline Male Hyptertension Relative Risk Function}
\usage{
hypertension_m_rr(x)
}
\arguments{
\item{x}{A vector of x values at which to evaluate the RR function}
}
\description{
Special Spline Male Hyptertension Relative Risk Function
}
