% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{ihd_rr}
\alias{ihd_rr}
\title{Get Special Fractional Polynomial Relative Risk Function (IHD extrapolated af
ter 100)}
\usage{
ihd_rr(B, extrapolation = TRUE)
}
\arguments{
\item{B}{The numeric vector of Beta values needed to produce a fractional poly
nomial}

\item{extrapolation}{Either TRUE(linear) or FALSE(capped) used for
extrapolating the RR after 150}
}
\description{
Get Special Fractional Polynomial Relative Risk Function (IHD extrapolated af
ter 100)
}
