% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_data_handling.R
\name{impute_with}
\alias{impute_with}
\title{What value is imputed when a variable is missing?}
\usage{
impute_with(var)
}
\arguments{
\item{var}{is a string -- a variable name.}
}
\value{
A string or integer that is appropriate for imputation.
}
\description{
In order to normalize code in future methods, missing data is removed either
through imputation when possible, or by throwing an error when the data is
necessary.  impute_with is called when a missing value \emph{can} be imputed.
}
