% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imputeWith}
\alias{imputeWith}
\title{Provides a value to impute with}
\usage{
imputeWith(var)
}
\arguments{
\item{var}{a variable name (as a character vector).}
}
\value{
An error if the variable is not suitable or imputation or the
imputation value if the variable is
}
\description{
Provides a value to impute with
}
