% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{intermahpr_base}
\alias{intermahpr_base}
\title{Takes Base InterMAHP inputs, returns formatted data}
\usage{
intermahpr_base(RelativeRisks = intermahpr::rr_default, PrevalenceConsumption,
  FemaleLightModerateBarrier, FemaleModerateHeavyBarrier, FemaleBingeBarrier,
  MaleLightModerateBarrier, MaleModerateHeavyBarrier, MaleBingeBarrier,
  UpperBound, Extrapolation, OutputPath = NULL, FilePrefix = "")
}
\arguments{
\item{RelativeRisks}{rr}

\item{PrevalenceConsumption}{pc}

\item{FemaleLightModerateBarrier}{flm}

\item{FemaleModerateHeavyBarrier}{fmh}

\item{FemaleBingeBarrier}{fbb}

\item{MaleLightModerateBarrier}{mlm}

\item{MaleModerateHeavyBarrier}{mmh}

\item{MaleBingeBarrier}{mbb}

\item{UpperBound}{ub}

\item{Extrapolation}{ext}

\item{OutputPath}{op}
}
\description{
Takes Base InterMAHP inputs, returns formatted data
}
