% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermahpr_constants.R
\docType{data}
\name{intermahpr_constants}
\alias{intermahpr_constants}
\title{Get and Set InterMAHPR environment/state variables}
\format{An object of class \code{environment} of length 2.}
\usage{
intermahpr_constants
}
\description{
Get and Set InterMAHPR environment/state variables
}
\keyword{datasets}
