% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{intermahpr_raw}
\alias{intermahpr_raw}
\title{Takes all possible inputs, computes specified AAFs, returns TMI}
\usage{
intermahpr_raw(pc = intermahpr::pc_default, rr = intermahpr::rr_default,
  ext = TRUE, lb = 0.03, ub = 250, bb = list(Female = 53.8, Male =
  67.25), gc = list(Female = 1.258^2, Male = 1.171^2), cb = list(Female =
  c(13.45, 26.9), Male = c(20.2, 40.4)))
}
\arguments{
\item{pc}{Prevalence / Consumption input}

\item{rr}{Relative Risk input}

\item{ext}{logical.  Answer to the question: "extrapolate linearly?"}

\item{lb}{Double, consumption lower bound, given in
units of grams ethanol per day}

\item{ub}{Double, consumption upper bound, given in
units of grams ethanol per day}

\item{bb}{Double vector, Binge consumption level, Gender stratified, given in
units of grams ethanol per day. Stratified by gender -- names(bb) must match
levels of rr$GENDER and pc$GENDER.}

\item{gc}{Gamma constant. The linear relationship between mean and standard
deviation within the gamma distribution that describes consumption among
current drinkers. Stratified by gender -- names(gc) must match levels of
rr$GENDER and pc$GENDER.}

\item{cb}{Consumption barriers. These define light/moderate/heavy drinking
groups. Given in units of grams ethanol per day. Stratified by gender --
names(cb) must match levels of rr$GENDER and pc$GENDER.}
}
\description{
The most granular AAF computation.  All inputs have default values, and levels
of Gender are inspected before any computations are done.
}
