% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intermahpr_sample_mm}
\alias{intermahpr_sample_mm}
\title{Default Morbidity and Mortality Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3360 rows and 7 columns.
}
\usage{
data(intermahpr_sample_mm)
}
\description{
Default morbidity and mortality data for Canada, 2007-2016.  Obtained from
CANSIM databases.
Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
