% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intermahpr_sample_pc}
\alias{intermahpr_sample_pc}
\title{Default Prevalence and Consumption Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 12 columns.}
\usage{
data(intermahpr_sample_pc)
}
\description{
Default prevalence and consumption data for Canada, 2007-2016.
Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
