% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intermahpr_sample_rr}
\alias{intermahpr_sample_rr}
\title{Default Relative Risk Curve Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67 rows and 24 columns.}
\usage{
data(intermahpr_sample_rr)
}
\description{
Default input to generate relative risk curves.  Standard formatting is
described in the InterMAHP user guides
}
\keyword{datasets}
