% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{intgrnd_factory}
\alias{intgrnd_factory}
\title{Factory for aaf integrands}
\usage{
intgrnd_factory(BB, LB, UB, R1, R2, LNXT_RR, BNGD_RR, N_GAMMA)
}
\arguments{
\item{BB}{dbl, binge barrier}

\item{LB}{dbl, lower bound of consumption}

\item{UB}{dbl, upper bound of consumption}

\item{R1}{dbl, proportion of drinkers below BB that do not binge}

\item{R2}{dbl, proportion of drinkers below BB that do binge}

\item{LNXT_RR}{fn, extrapolated relative risk for nonbingers}

\item{BNGD_RR}{fn, extrapolated relative risk for bingers}

\item{N_GAMMA}{fn, normalized distribution of drinkers}
}
\value{
a function object that represents the generalized integrand found in
 the generalized AAF computation outlined in the the InterMAHP user guide
}
\description{
Given function data (gamma distribution specs, relative risk
 curves, binge ratios) integrand_factory produces a function for use as an
 integrand in AAF computations
}
