% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{linear_extrapolation_factory}
\alias{linear_extrapolation_factory}
\title{Produce an extrapolation-modified relative risk curve}
\usage{
linear_extrapolation_factory(BASE_RR, X2, Y2, SLOPE)
}
\arguments{
\item{BASE_RR}{continuous function in x as returned by base_rr_factory}

\item{X2}{extrapolate after x = X2}

\item{Y2}{Y2 = BASE_RR(X2)}

\item{SLOPE}{slope of the extrapolation}
}
\value{
a function whose values above 150 (100 for IHD) are linearly
 extrapolated (see InterMAHP guide for details)
}
\description{
given a relative risk curve and some metadata, produces a
 risk curve that has the desired extrapolation behaviour beyond 150 g/day,
 or 100 g/day for IHD.
}
