% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lowerVars}
\alias{lowerVars}
\title{Converts data variable names to lower case and returns data}
\usage{
lowerVars(.data)
}
\arguments{
\item{.data}{a data_frame object}
}
\value{
The dataset with variable names in lower case
}
\description{
Lowers the case of all variable names.  For compatibility with SAS version.
}
