% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeAcutePancreatitisRisk}
\alias{makeAcutePancreatitisRisk}
\title{Factory for Special Spline Female Acute Pancreatitis Relative Risk Functions}
\usage{
makeAcutePancreatitisRisk()
}
\description{
Factory for Special Spline Female Acute Pancreatitis Relative Risk Functions
}
