% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeBaseRisk}
\alias{makeBaseRisk}
\title{Factory for base well-defined rel. risk}
\usage{
makeBaseRisk(im, gender, form, betas)
}
\arguments{
\item{im}{string, Intermahp condition code}

\item{gender}{string}

\item{form}{string, "FP", "Step", or "Spline"}

\item{betas}{dbl, vector of length 16}
}
\value{
a function object that describes a base relative risk curve
}
\description{
Choose which relative risk function is appropriate from the given input
}
