% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeBingeRisk}
\alias{makeBingeRisk}
\title{Factory for binge-modified well-defined rel. risk.}
\usage{
makeBingeRisk(im, bingef, ext_risk)
}
\arguments{
\item{im}{string, Intermahp condition code}

\item{bingef}{double, rescales injuries}

\item{ext_risk}{continuous vector valued function as returned by lin_ext_fact}
}
\value{
a function whose values are binge modified.  This affects conditions
 5.2 and 5.5, ischaemic heart disease and stroke respectively, by removing a
 protective effect at low levels of consumption
}
\description{
Produces a Relative-Risk-For-Bingers curve from the given input
Given a list that contains the string variable IM and the double
 variable BINGEF, produces a Relative-Risk-For-Bingers curve.  IM is used for
 the curves for IHD and ischaemic stroke, where any protective J-shape is
 forfeited by bingers.  All curves are then multiplied by BINGEF which has
 the value of 1 in all cases except for condition classes 7, 8, 9, which
 account for motor vehicle collisions, unintentional injuries, and
 intentional injuries.
}
