% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-calibrated-aaf-cmp.R
\name{makeConditionalProbability}
\alias{makeConditionalProbability}
\title{Factory for loglinear conditional probability functions}
\usage{
makeConditionalProbability(slope, lb)
}
\value{
Conditional probability mass function for risk incurred as a result of
exposure
}
\description{
Invokes slope_calibration and wraps the result in a loglinear
probability mass function
}
