% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-calibrated-aaf-cmp.R
\name{makeCurrentCalibratedFactory}
\alias{makeCurrentCalibratedFactory}
\title{Factory for current drinker's AAF computer factory for a condition with
 calibrated incidence estimator}
\usage{
makeCurrentCalibratedFactory(target, clbr_mass, lb, ub)
}
\description{
Factory for current drinker's AAF computer factory for a condition with
 calibrated incidence estimator
}
