% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-free-aaf-cmp.R
\name{makeCurrentFreeComponentFactory}
\alias{makeCurrentFreeComponentFactory}
\title{Factory for the current drinker's component in an AAF computer factory for a
 condition with well-defined relative risk}
\usage{
makeCurrentFreeComponentFactory(ext_risk, binge_risk)
}
\description{
Factory for the current drinker's component in an AAF computer factory for a
 condition with well-defined relative risk
}
