% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-free-aaf-cmp.R
\name{makeCurrentFreeFactory}
\alias{makeCurrentFreeFactory}
\title{Factory for current drinker's AAF computer factory for a condition with
well-defined relative risk}
\usage{
makeCurrentFreeFactory(ext_risk, binge_risk, rr_fd)
}
\description{
Exported for use in shiny app
}
