% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeExtrapolatedRisk}
\alias{makeExtrapolatedRisk}
\title{Factory for linearly extrapolated well-defined rel. risk}
\usage{
makeExtrapolatedRisk(base_risk, x2, y2, slope)
}
\arguments{
\item{base_risk}{continuous function in x as returned by makeBaseRisk}

\item{x2}{extrapolate after x = x2}

\item{y2}{y2 = base_risk(x2)}

\item{slope}{slope of the extrapolation}
}
\value{
a function whose values after x2 are are linearly extrapolated
 (see InterMAHP guide for details)
}
\description{
Produces an extrapolation-modified relative risk curve.
 Given a relative risk curve and some metadata, produces a
 risk curve that has the desired extrapolation behaviour beyond 150 g/day,
 or 100 g/day for IHD.
}
