% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-calibrated-aaf-cmp.R
\name{makeFormerCalibratedFactory}
\alias{makeFormerCalibratedFactory}
\title{Factory for former drinker's AAF computer factory for a condition with
 calibrated incidence estimator}
\usage{
makeFormerCalibratedFactory(...)
}
\description{
The relative risk for former drinkers is 1 by definition, so these functions
 are identically zero.
}
