% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-free-aaf-cmp.R
\name{makeFormerFreeComponentFactory}
\alias{makeFormerFreeComponentFactory}
\title{Factory for the former drinker's component in an AAF computer factory for a
 condition with well-defined relative risk}
\usage{
makeFormerFreeComponentFactory(rr_fd)
}
\description{
Factory for the former drinker's component in an AAF computer factory for a
 condition with well-defined relative risk
}
