% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-free-aaf-cmp.R
\name{makeFormerFreeFactory}
\alias{makeFormerFreeFactory}
\title{Factory for former drinker's AAF computer factory for a condition with
well-defined relative risk}
\usage{
makeFormerFreeFactory(ext_risk, binge_risk, rr_fd)
}
\description{
Exported for use in shiny app
}
