% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeFractionalPolynomial}
\alias{makeFractionalPolynomial}
\title{Factory for Fractional Polynomial Relative Risk Functions}
\usage{
makeFractionalPolynomial(betas)
}
\arguments{
\item{betas}{The numeric vector of Beta values needed to produce a fractional
polynomial}
}
\description{
Factory for Fractional Polynomial Relative Risk Functions
}
