% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rr.R
\name{makeFreeFactories}
\alias{makeFreeFactories}
\title{Factory for AAF computer factories: conditions with well-defined rel. risk}
\usage{
makeFreeFactories(.data)
}
\description{
Exported for use in shiny app
}
