% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-exposure.R
\name{makeGamma}
\alias{makeGamma}
\title{Factory for gamma distributions}
\usage{
makeGamma(shape, scale)
}
\arguments{
\item{shape, scale}{Gamma parameters for dgamma}
}
\description{
We want easier access to gamma distributions with a wide array of fixed params
}
