% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeHivRisk}
\alias{makeHivRisk}
\title{Factory for Special HIV Relative Risk Functions}
\usage{
makeHivRisk(gender)
}
\arguments{
\item{gender}{decides which function to return}
}
\description{
Factory for Special HIV Relative Risk Functions
}
