% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-rr.R
\name{makeHypertensionRisk}
\alias{makeHypertensionRisk}
\title{Factory for Special Spline Hypertension Relative Risk Functions}
\usage{
makeHypertensionRisk(gender)
}
\arguments{
\item{gender}{decides which function to return}
}
\description{
Factory for Special Spline Hypertension Relative Risk Functions
}
