% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-exposure.R
\name{makeNormalizedGamma}
\alias{makeNormalizedGamma}
\title{Factory for normalized gamma distributions}
\usage{
makeNormalizedGamma(shape, scale, factor)
}
\arguments{
\item{shape, scale}{Gamma parameters for dgamma}

\item{factor}{Factor to scale distribution by}
}
\description{
Factory for normalized gamma distributions
}
