% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories-free-aaf-cmp.R
\name{makePreventableFraction}
\alias{makePreventableFraction}
\title{Factory for preventable fraction functions}
\usage{
makePreventableFraction(bb, non_bingers, bingers, ext_risk, binge_risk)
}
\arguments{
\item{bb}{dbl, binge barrier}

\item{non_bingers}{dbl, proportion of drinkers below BB that do not binge}

\item{bingers}{dbl, proportion of drinkers below BB that do binge}

\item{ext_risk}{fn, extrapolated relative risk for nonbingers}

\item{binge_risk}{fn, extrapolated relative risk for bingers}
}
\description{
Produces the combined and scaled function that respresents the
 preventable fraction of disease that, when integrated against exposure,
 produces an attributable fraction term.
}
