% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeProduct}
\alias{makeProduct}
\alias{\%prod\%}
\title{Factory for Pointwise Function Products}
\usage{
makeProduct(f, g)

f \%prod\% g
}
\arguments{
\item{f, g}{function that takes a single argument and produces a value that is
a valid argument for the `*` function}
}
\description{
factory that produces the product of a pair of functions, where
the product used is pointwise multiplication

binary operator for product_factory
}
\section{Functions}{
\itemize{
\item \code{\%prod\%}: 
}}

