% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{name_cuts}
\alias{name_cuts}
\title{Postprocessing function that extracts Light, Moderate, Heavy group AAFs}
\usage{
name_cuts(aaf_table)
}
\description{
Postprocessing function that extracts Light, Moderate, Heavy group AAFs
}
