% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{normalized_gamma_factory}
\alias{normalized_gamma_factory}
\title{Factory for normalized gamma distributions}
\usage{
normalized_gamma_factory(pc_specs)
}
\arguments{
\item{pc_specs}{is a tibble that contains the variables:
GAMMA_SHAPE: dbl
GAMMA_SCALE: dbl
DF: dbl}
}
\value{
a function object that repreents the normalized gamma distribution
 (defined as defined as DF*gamma(x,shape,scale))
}
\description{
Factory for normalized gamma distributions
}
