% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_bc}
\alias{pc_bc}
\title{Default Prevalence and Consumption Data restricted to BC}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 12 columns.}
\usage{
data(pc_bc)
}
\description{
Default prevalence and consumption data from BC, 2015.
Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
