% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g00-default-data.R
\docType{data}
\name{pc_bc_can}
\alias{pc_bc_can}
\title{Default Prevalence and Consumption Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 12 columns.}
\usage{
data(pc_bc_can)
}
\description{
Default prevalence and consumption data from BC and Canada, 2015.
Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
