% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pc.R
\docType{data}
\name{pc_constants}
\alias{pc_constants}
\title{List of pc constants}
\format{An object of class \code{character} of length 3.}
\usage{
pc_constants
}
\description{
List of pc constants
}
\keyword{datasets}
