% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R, R/derive_params_from_pc.R
\docType{data}
\name{pc_default}
\alias{pc_default}
\alias{pc_default}
\title{Default Prevalence and Consumption Data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 12 rows and 12 columns.}
\usage{
data(pcr_default)

data(pcr_default)
}
\description{
Default prevalence and consumption data from BC and Canada, 2015.input to
generate relative risk curves.  Standard formatting is described in the
InterMAHP user guides

Default prevalence and consumption data from BC and Canada, 2015.input to
generate relative risk curves.  Standard formatting is described in the
InterMAHP user guides
}
\keyword{datasets}
