% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pc.R
\docType{data}
\name{pc_vars}
\alias{pc_vars}
\title{List of variables expected to be in an input PC sheet}
\format{An object of class \code{character} of length 12.}
\usage{
pc_vars
}
\description{
List of variables expected to be in an input PC sheet
}
\keyword{datasets}
