% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{preventable_fraction_factory}
\alias{preventable_fraction_factory}
\title{Factory for preventable fractions}
\usage{
preventable_fraction_factory(BB, R1, R2, LNXT_RR, BNGD_RR)
}
\arguments{
\item{BB}{dbl, binge barrier}

\item{R1}{dbl, proportion of drinkers below BB that do not binge}

\item{R2}{dbl, proportion of drinkers below BB that do binge}

\item{LNXT_RR}{fn, extrapolated relative risk for nonbingers}

\item{BNGD_RR}{fn, extrapolated relative risk for bingers}
}
\value{
a function object valid on the domain [LB,UB]
}
\description{
Produces the combined and scaled function that respresents the
 preventable fraction of disease that, when integrated against exposure,
 produces an attributable fraction term.
}
