% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02-process.R
\name{process_dh_aaf}
\alias{process_dh_aaf}
\title{Joins DH and AAF data, fills in missing aaf_cmps}
\usage{
process_dh_aaf(dh, aaf)
}
\arguments{
\item{dh}{death/hosp tibble as produced by derive_dh}

\item{aaf}{aaf tibble as produced by base_aafs}
}
\description{
Provides continuous AAF computing functions for 100% attributable conditions
either by calibrating from counts and drinkers (3.2|4.[123]|5.3|6.[15]) or
rescaling similar aaf_cmps (4.[467]|8.5|9.2).
Granulates [89].all into respective boxes, copies 6.2 into 5.7.
}
