% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02-process.R
\name{process_dh_pc}
\alias{process_dh_pc}
\title{Joins DH and PC data}
\usage{
process_dh_pc(dh, pc)
}
\arguments{
\item{dh}{Deaths/Hosps as returned by format_dh}

\item{pc}{Prev/Cons as returned by derive_pc}
}
\description{
Joins DH and PC data
}
