% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{product_factory}
\alias{product_factory}
\title{Pointwise Function Product Factory}
\usage{
product_factory(f, g)
}
\arguments{
\item{f, g}{function that takes a single argument and produces a value that is
a valid argument for the `*` function}
}
\description{
factory that produces the product of a pair of functions, where
the product used is pointwise multiplication
}
