% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{queryYoung}
\alias{queryYoung}
\title{Get age-groups present in morbidity/mortality dataset but absent in prev-cons
dataset.}
\usage{
queryYoung(.data)
}
\description{
Get age-groups present in morbidity/mortality dataset but absent in prev-cons
dataset.
}
