% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\name{reduce_pc}
\alias{reduce_pc}
\title{Reduces the per capita consumption and binge drinker prevalence}
\usage{
reduce_pc(.data, reduction)
}
\arguments{
\item{.data}{a prev/cons sheet as would be accepted by format_pc}

\item{reduction}{a percentage of the current consumption expected in the
scenario under study}
}
\description{
Reduces the per capita consumption and binge drinker prevalence
}
