% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pc.R
\name{rescale}
\alias{rescale}
\title{Rescale Population Data}
\usage{
rescale(.data, scale = 1)
}
\arguments{
\item{.data}{cleaned population data with constants set}

\item{scale}{a percentage of the current consumption expected in the
scenario under study}
}
\description{
Rescale Population Data
}
