% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{rr_chooser}
\alias{rr_chooser}
\title{Relative Risk Function Chooser}
\usage{
rr_chooser(row)
}
\arguments{
\item{row}{is a row from the Relative Risk input frame plus an extrapolation
column.
The necessary structure for each row is
IM > Condition > Gender > Outcome > RR_FD > BingeF > Function > B1-B16
> extrapolation}
}
\value{
A list of:
           1. list: IM, Condition Name, Gender, Type (Morb, Mort, Comb)
           2. RR_FD constant
           3. Relative risk function
           4. Relative risk for bingers
}
\description{
Given a row with the standard Relative Risk input structure (plus extrapolation T/F column),
             determines a relative risk function and a relative risk for binge-drinkers function to be returned.
}
