% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\docType{data}
\name{rr_default}
\alias{rr_default}
\title{Default Relative Risk Curve Defintion}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 54 rows and 23 columns.}
\usage{
data(rr_default)
}
\description{
Default input to generate relative risk curves.  Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
