% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rr_master}
\alias{rr_master}
\title{Master rr list for Version 2 and beyond}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 24 columns.}
\usage{
data(rr_master)
}
\description{
Master rr list for Version 2 and beyond
}
\keyword{datasets}
