% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rr_roerecke}
\alias{rr_roerecke}
\title{Default Relative Risk Curves with Roerecke IHD treatment}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 23 columns.}
\usage{
data(rr_roerecke)
}
\description{
Default input to generate relative risk curves, uses Roerecke IHD treatment.
Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
