% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rr.R
\docType{data}
\name{rr_vars}
\alias{rr_vars}
\title{List of variables expected to be in an RR sheet}
\format{An object of class \code{character} of length 24.}
\usage{
rr_vars
}
\description{
List of variables expected to be in an RR sheet
}
\keyword{datasets}
