% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data_declarations.R
\docType{data}
\name{rr_zhao}
\alias{rr_zhao}
\title{Deafult Relative Risk Curves with Zhao IHD treatment}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 23 columns.}
\usage{
data(rr_zhao)
}
\description{
Default input to generate relative risk curves, uses Zhao IHD treatment.
Standard formatting is described in the InterMAHP user guides
}
\keyword{datasets}
