% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01-wrangle.R
\name{scale_pc}
\alias{scale_pc}
\title{Scales the per capita consumption and binge drinker prevalence}
\usage{
scale_pc(.data, scale = 1)
}
\arguments{
\item{.data}{a prev/cons sheet as would be accepted by format_pc}

\item{scale}{a percentage of the current consumption expected in the
scenario under study}
}
\description{
Scales the per capita consumption and binge drinker prevalence
}
