% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02_aaf_factory.R
\name{scaled_aaf_cmp_factory}
\alias{scaled_aaf_cmp_factory}
\title{Rescales a given AAF computing function into a wholly attributable AAF
computing function}
\usage{
scaled_aaf_cmp_factory(aaf_specs)
}
\arguments{
\item{aaf_specs}{is a tibble that contains the variables:
AAF_CMP: fn
AAF_TOTAL: dbl
P_FD: dbl
RR_FD: dbl
INTGRND: fn}
}
\description{
When a 100% condition is reasonably similar to a partially attributable
condition, it is reasonable to simply rescale the partially attributable AAF
computing function by the reciprocal of AAF_TOTAL and attribute that to the
100% condition.
}
