% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pc.R
\name{setPopnConstants}
\alias{setPopnConstants}
\title{Set Population Constants}
\usage{
setPopnConstants(.data, bb = list(Female = 53.8, Male = 67.25), lb = 0.03,
  ub = 250)
}
\arguments{
\item{bb}{binge barrier}

\item{lb}{lower bound of consumption}

\item{ub}{upper bound of consumption}
}
\description{
Set Population Constants
}
