% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02-factories.R
\name{slope_calibration}
\alias{slope_calibration}
\title{Slope Factory}
\usage{
slope_calibration(IM, COUNT, DRINKERS, N_GAMMA, LB, BB, UB)
}
\arguments{
\item{IM}{InterMAHP condition code}

\item{COUNT}{Count of condition events over the relevant time period}

\item{DRINKERS}{Estimated # of drinkers in population}

\item{N_GAMMA}{Normalized gamma distribution (used as exposure mass function)}

\item{LB, BB, UB}{lower, binge, and upper bounds of consumption}
}
\value{
slope of loglinear conditional probability mass function for risk as a
result of exposure
}
\description{
Given a list that contains the true count among a given cohort of
a wholly alocohol-attributable condition, the number of drinkers in that
cohort, the conditions's IM code, and the gamma function detailing consumption
among the cohort, this function calibrates a slope for a loglinear function
that estimates the conditional probablity of developing the given condition.
The function should be interpreted as conditional probability mass.  I.e., the
integral of the function over the relevant range is the probability that a
drinker will be afflicted by an event of the given condition over the period
of time that the given Count variable was collected.

Uses a nonlinear optimizer (COBYLA) to find a loglinear slope for the
function f(x) = max(1, exp(k(x-t))) that mimizes the difference between
integral(N_GAMMA * (f-1), 0.03, UB) and yearly prevalence (count/drinkers).

The goal is to produce a continuous analogue to the relative risk curve for
conditions that are wholly attributable to alcohol. The assumption is made
that such a condition has a loglinear thresholded (i.e. f(x)=1 for x<t)
conditional probablity function on the interval of concern (0.03 to UB grams
of ethanol/day, averaged over 1yr).

This conditional probability is used to portion a 1.00 AAF_TOTAL among the
drinking population.

The nonlinear optimization always estimates conservatively.  For conditions
with observed incidence at leats 1.8 per 10,000 drinkers, the AAF will round
to 1 within 2 decimal places.  There is no such guarantee for conditions with
incidence less than 1.8 per 10,000 drinkers.
}
