% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02-postprocessors.R
\name{split_outcome}
\alias{split_outcome}
\title{Postprocessing function that tidies base AAF by outcome type}
\usage{
split_outcome(aaf_table)
}
\arguments{
\item{aaf_table}{Any table with an OUTCOME variable whose levels are:
"Mortality", "Morbidity", and "Combined".}
}
\value{
aaf_table with an OUTCOME variable whose levels are: "Mortality" and
 "Morbidity" and 2c+t+b rows, where c is the number of OUTCOME = "Combined"
 rows in aaf_table, t is the number of OUTCOME = "Mortaility", and b is the
 number of OUTCOME  = "Morbidity".
}
\description{
Splits "Combined" levels of an OUTCOME variable into Morbidity and Mortality
}
