% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{split_sim}
\alias{split_sim}
\title{Split Similar AAF rows}
\usage{
split_sim(aaf_table)
}
\arguments{
\item{aaf_table}{as produced by name_cuts (i.e has names(aaf_table) of
REGION, YEAR, GENDER, AGE_GROUP, IM, CONDITION, OUTCOME,
AAF_FD, AAF_LD, AAF_MD, AAF_HD, AAF_TOTAL)}
}
\description{
condition specific method.
  - 4.5 AAFs are proportionally distributed to 4.4, 4.6, and 4.7.
  - 5.3 receives the AAF distribution of LD: 0, MD: 0, HD: 0.
  - 8.all and 9.all are split exactly over 8.1-4, 8.6 and 9.1, 9.3-5 resp.,
      and proportionally over 8.5 and 9.2 resp.
}
