% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{vintegrate}
\alias{vintegrate}
\title{Postprocessing helper function that integrates a list of functions from a
list of lower bounds to a list of upper bounds.}
\usage{
vintegrate(funs, vlower, vupper)
}
\description{
Integrate that takes as input a list funs of functions, a vector vlower of
"lower" values and a vector vupper of "upper" values, assumes that
length(vlower) = length(vupper) = length(funs), and returns a
vector of integrals of funs between the lower and upper values.
}
\details{
This function is not optimized or "vectorized," it simply assigns memory and
uses a for loop to perform each integral.
}
