% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s00_data_handling.R
\docType{data}
\name{zero}
\alias{zero}
\title{Dummy function for allocating memory}
\format{An object of class \code{list} of length 1.}
\usage{
zero
}
\arguments{
\item{...}{Accepts any input}
}
\value{
Always returns 0
}
\description{
Dummy function for allocating memory
}
\keyword{datasets}
