% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02_aaf_factory.R
\name{aaf_cmp_factory}
\alias{aaf_cmp_factory}
\title{Factory for alcohol attributable fraction computing functions}
\usage{
aaf_cmp_factory(aaf_specs)
}
\arguments{
\item{aaf_specs}{is a tibble that contains the variables:
LB: dbl
UB: dbl
RR_FD: dbl
P_FD: dbl
INTGRND: fn}
}
\value{
a function object that takes as input a vector x of values >= 0.03 and
 returns the alcohol attributable fraction between 0.03 and x for outcome
 OUTCOME due to condition CONDITION among cohort GENDER * AGE_GROUP in the
 place and time REGION * YEAR
}
\description{
Given function data (gamma distribution specs, relative risk
 curves, binge ratios) aaf_factory produces a function f that computes aaf
 aafs from a given interval
}
