% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{aaf_fd}
\alias{aaf_fd}
\title{Computes alcohol attributable fraction for former drinkers}
\usage{
aaf_fd(LB, UB, RR_FD, P_FD, INTGRND)
}
\arguments{
\item{LB}{dbl, lower bound of consumption}

\item{UB}{dbl, upper bound of consumption}

\item{RR_FD}{dbl, relative risk for former drinkers}

\item{P_FD}{dbl, proportion of former drinkers}

\item{INTGRND}{fn, continuous function to be integrated}
}
\value{
dbl<1 to be interpreted as a fraction
}
\description{
Given risk, population, and current drinker integrand information
 aaf_fd computes the alcohol attributable fraction among former drinkers
}
