% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{acute_pancreatitis_f_rr}
\alias{acute_pancreatitis_f_rr}
\title{Special Spline Female Acute Pancreatitis Relative Risk Function}
\usage{
acute_pancreatitis_f_rr(x)
}
\arguments{
\item{x}{A vector of x values at which to evaluate the RR function


0.04 = 1/25
0.0007304602 = 1/(37^2)}
}
\description{
Special Spline Female Acute Pancreatitis Relative Risk Function
}
