% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{addCurrentFraction}
\alias{addCurrentFraction}
\title{Compute a given scenario's AAF for current drinkers and adds it to the
scenario}
\usage{
addCurrentFraction(.data, var_name = "aaf_cd")
}
\description{
Compute a given scenario's AAF for current drinkers and adds it to the
scenario
}
