% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{addTotalFraction}
\alias{addTotalFraction}
\title{Compute a given scenario's Total AAF and adds it to the scenario}
\usage{
addTotalFraction(.data, var_name = "aaf")
}
\description{
Compute a given scenario's Total AAF and adds it to the scenario
}
