% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_process_pc_rr.R
\name{assemble}
\alias{assemble}
\title{Collect and assemble AAF data from formatted RR and PC data}
\usage{
assemble(pc, rr, ext, lb, ub, bb, gc)
}
\arguments{
\item{pc}{Prevalence / Consumption input as produced by format_pc}

\item{rr}{Relative Risk input as produced by format_rr}

\item{ext}{logical, extrapolate linearly?}

\item{lb}{Double, consumption lower bound}

\item{ub}{Double, consumption upper bound}

\item{bb}{Double vector, Binge consumption level, Gender stratified}

\item{gc}{Gamma constant.  The linear relationship between mean and standard
deviation within the gamma distribution that describes consumption among
current drinkers.  Stratified by gender -- names(gc) must match levels of
rr$GENDER and pc$GENDER.}
}
\description{
Collect and assemble AAF data from formatted RR and PC data
}
