% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_aaf_evaluation.R
\name{compute_aafs}
\alias{compute_aafs}
\title{Compute AAFs for all conditions as separated by the given cutpoints}
\usage{
compute_aafs(aaf_table, cuts)
}
\arguments{
\item{aaf_table}{A tibble as returned by assemble (i.e. has the following vars
GENDER: chr, gender levels
AAF_CMP: fn, VIVO, computes AAF from LB to each input
AAF_FD: dbl, AAF for former drinkers
LB: dbl, consumption lower bound
UB: dbl, consumption upper bound
)}

\item{cuts}{A sorted list of double vectors indexed by aaf_table$GENDER st.
each value is between LB and UB}
}
\description{
Compute AAFs for all conditions as separated by the given cutpoints
}
