% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{computeCurrentFraction}
\alias{computeCurrentFraction}
\title{Computes a given scenario's AAF for current drinkers}
\usage{
computeCurrentFraction(.data)
}
\description{
Computes a given scenario's AAF for current drinkers
}
