% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{computeFormerFraction}
\alias{computeFormerFraction}
\title{Compute a given scenario's AAF for former drinkers}
\usage{
computeFormerFraction(.data)
}
\description{
Compute a given scenario's AAF for former drinkers
}
