% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01-wrangle.R
\name{derive_rr}
\alias{derive_rr}
\title{Derives relative risk curves from relative risk data}
\usage{
derive_rr(.data, ext = TRUE)
}
\arguments{
\item{.data}{tibble as formatted in format_rr}

\item{ext}{logical indicator of extrapolation, (TRUE = linear, FALSE = capped)}
}
\description{
Derives relative risk curves from relative risk data
}
