% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{derive_v0_pc}
\alias{derive_v0_pc}
\title{Derives probability distribution parameters from Prevalence and Consumption
data}
\usage{
derive_v0_pc(pc, bb = c(53.8, 67.25), lb = c(0.03, 0.03), ub = c(250,
  250))
}
\arguments{
\item{bb}{User supplied gender stratified binge barrier}

\item{lb}{Constants used:
0.002739726 = 1/365
0.7893 = unit conversion (millilitres to grams) for ethanol}

\item{PC}{is assumed type tibble with names(PC) = c(YEAR, REGION, GENDER
, AGE_GROUP, POPULATION, PCC_LITRES_YEAR, CORRECTION_FACTOR,
RELATIVE_CONSUMPTION, P_LA, P_FD, P_CD, P_BD).
This function is internally called only, not exported.

Note that the bundled data set pc_default satisfies these constraints.}
}
\value{
A tibble with additional variables:
  PCC_G_DAY: Per capita consumption, converted from litres/year to grams/day
  DRINKERS: Total drinkers in region, computed as POPULATION*P_CD
  PCAD: Average consumption among drinkers
  PCC_AMONG_DRINKERS: Computed from PCAD and relative consumption
  GAMMA_CONSTANT: The distribution of drinkers at drinking level of x grams
    per day is fit by a gamma distribution whose mean and standard deviation
    are related by GAMMA_CONSTANT, stratified by gender
  GAMMA_SHAPE: Computed from GAMMA_CONSTANT
  GAMMA_SCALE: Computed from GAMMA_CONSTANT
  INDEX: Among vector inputs, which index refers to this cohort?
  BB: Binge barrier
  LB: Lower bound of integration (0.03)
  UB: Upper bound of integration (user-specified, typical values are 150,250)
  NC: Normalizing constant for the gamma distribution (how much of the gamma
  distribution is between LB and UB?)
  DF: Deflation factor for the gamma distribution (DF = P_CD/NC)
  P_BAT: Proportion of bingers above threshold (BB)
  R1, R2: Ratios computed at this step and used in AAF computations
}
\description{
From a given prevalence and consumption dataframe (or similar), computes re-
-quired data for AAF computation including Per Capita Consumption among
drinkers, and Gamma Shape/Scale parameters.  Also computes constants
necessary for AAF integration, namely the Deflation factor and Binge-Split
ratios.
}
\details{
Gamma parameters are derivable entirely from prevalence and consumption,
no additional input is necessary.
}
