% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{extract_prevcons}
\alias{extract_prevcons}
\title{Helper function that tidies base AAF into the relevant prevalence and
consumption data}
\usage{
extract_prevcons(aaf_table)
}
\description{
Helper function that tidies base AAF into the relevant prevalence and
consumption data
}
