% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{format_v0_pc}
\alias{format_v0_pc}
\title{Format prevalence and consumption input data to our desired specifications}
\usage{
format_v0_pc(pc)
}
\description{
expected variables:
 Needed:
   Proportion, Former Drinkers
   Proportion, Binge Drinkers
   Proportion, Lifetime Abstainers
   Proportion, Current Drinkers
   Gamma Distribution Parameters:
     Per Capita Consumption
     Population
     Relative Consumption
     Correction Factor
   Gender (in order to pair with RR curves)
 Imputable as "Unspecified":
   Region
   Year
   Gender
   Age Group
}
