% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{format_v0_rr}
\alias{format_v0_rr}
\title{Format relative risk input data to our desired specifications}
\usage{
format_v0_rr(rr)
}
\arguments{
\item{rr}{is a dataset that can be converted to tibble with the expected
variables}
}
\description{
Formats all variable names to upper case and imputes imputable missing data.
Intended for a relative risk sheet as described in the intermahp user guide.

expected variables:
Needed:
 Gender (in order to pair with PC data)
 Function (FP, Spline, with valid IM, or Step, for HIV only)
 B1-B16 (only if Function == FP. need nonzero betas, rest can be missing)
Imputable:
 IM
 Condition
 Outcome
 RR_FD
 BingeF
}
