% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{fp_rr}
\alias{fp_rr}
\title{Get Fractional Polynomial Relative Risk Function}
\usage{
fp_rr(B, extrapolation = TRUE, extrapolate_using = c(100, 150))
}
\arguments{
\item{B}{The numeric vector of Beta values needed to produce a fractional poly
nomial}

\item{extrapolation}{Either TRUE(linear) or FALSE(capped) used for
extrapolating the RR after 150}
}
\description{
Get Fractional Polynomial Relative Risk Function
}
