% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%prod\%}
\alias{\%prod\%}
\title{Binary operator for Pointwise Function Products}
\usage{
f \%prod\% g
}
\arguments{
\item{f, g}{function that takes a single argument and produces a value that is
a valid argument for the `*` function}
}
\description{
binary operator for product_factory
}
