% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxx_factories.R
\name{hypertension_f_rr}
\alias{hypertension_f_rr}
\title{Special Spline Female Hypertension Relative Risk Function}
\usage{
hypertension_f_rr(x)
}
\arguments{
\item{x}{A vector of x values at which to evaluate the RR function


0.0025 = 1/400}
}
\description{
Special Spline Female Hypertension Relative Risk Function
}
