% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01_relative_risk_curves.R
\name{hypertension_m_rr}
\alias{hypertension_m_rr}
\title{Special Spline Male Hyptertension Relative Risk Function}
\usage{
hypertension_m_rr(x)
}
\arguments{
\item{x}{A vector of x values at which to evaluate the RR function

0.01851852 = 1/54
0.0001777778 = 1/(75^2)}
}
\description{
Special Spline Male Hyptertension Relative Risk Function
}
