% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaf_factory.R
\name{intgrnd_factory}
\alias{intgrnd_factory}
\title{Factory for aaf integrands}
\usage{
intgrnd_factory(aaf_specs)
}
\arguments{
\item{aaf_specs}{is a tibble that contains the variables:
BB: dbl
LB: dbl
UB: dbl
R1: dbl
R2: dbl
LNXT_RR: fn
BNGD_RR: fn
N_GAMMA: fn}
}
\value{
a function object that represents the generalized integrand found in
 the generalized AAF computation outlined in the the InterMAHP user guide
}
\description{
Given function data (gamma distribution specs, relative risk
 curves, binge ratios) integrand_factory produces a function for use as an
 integrand in AAF computations
}
