% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_process_pc_rr.R
\name{join_pc_rr}
\alias{join_pc_rr}
\title{Joins PC and RR and obtains AAF functions}
\usage{
join_pc_rr(pc, rr)
}
\arguments{
\item{RR}{relative risk tibble as produced by derive_v*_rr}

\item{PC}{prevalence and consumption tibble as produced by derive_v*_pc}
}
\value{
tibble with one row per unique region.year.gender.age_group.im combn,
 an AAF_FD variable with alc.attr. fraction for former drinkers, and an
 AAF_CMP variable that
}
\description{
Performs a full_join of RR and PC over GENDER and then obtains an
 AAF function and AAF_FD computation for each observation.  Does a sanity
 check, ensuring that the joining variable is well-matched.
}
