% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rr.R
\name{makeCalibratedFactories}
\alias{makeCalibratedFactories}
\title{Factory for AAF computer factories: conditions requiring calibration against
population statistics}
\usage{
makeCalibratedFactories(rr, pc, dh)
}
\description{
Factory for AAF computer factories: conditions requiring calibration against
population statistics
}
