% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-factories.R
\name{normalized_gamma_factory}
\alias{normalized_gamma_factory}
\title{Factory for normalized gamma distributions}
\usage{
normalized_gamma_factory(GAMMA_SHAPE, GAMMA_SCALE, DF)
}
\arguments{
\item{GAMMA}{Input gamma distribution}

\item{FACTOR}{Factor to scale GAMMA by}
}
\description{
Factory for normalized gamma distributions
}
