% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{outcome_splitter}
\alias{outcome_splitter}
\title{Helper function that tidies base AAF into a certain outcome type}
\usage{
outcome_splitter(aaf_table, outcome)
}
\description{
Filters AAFs into any of mortality, morbidity, or combined.  Exported for
 use in the Shiny app, called behind the scenes when invoking intermahp_base.
}
