% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{outcome_splitter}
\alias{outcome_splitter}
\title{Helper function that tidies base AAF into a certain outcome type}
\usage{
outcome_splitter(aaf_table, outcome)
}
\description{
Helper function that tidies base AAF into a certain outcome type
}
