% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{set_rr}
\alias{set_rr}
\title{Choose which relative risk function is appropriate from the given input}
\usage{
set_rr(rr_specs)
}
\arguments{
\item{rr_specs}{a list that contains the string variables IM, GENDER, and
FUNCTION, and sixteen double variables B1 through B16}
}
\value{
a function object that describes a base relative risk curve
}
\description{
Given a list that contains the string variables IM, GENDER, and
 FUNCTION, and sixteen double variables B1 through B16, returns a relative
 risk function.
 Currently supported levels of IM can be found in the User Guide. Supported
 levels for GENDER are Female and Male, and for FUNCTION are FP, Spline, and
 Step.
 IM, FUNCTION, and GENDER are used to determine which function to produce in
 extraordinary cases.  The most common case, however, is the FP (Fractional
 Polynomial) model. When FUNCTION = FP and the given condition is not
 extraordinary, a relative risk function will still be produced even if
 Gender/IM are imputed by intermahpr.
}
