% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_curves.R
\name{simple_rr}
\alias{simple_rr}
\title{Get Simple Fractional Polynomial Relative Risk Function}
\usage{
simple_rr(B, extrapolation = TRUE)
}
\arguments{
\item{B}{The numeric vector of Beta values needed to produce a fractional polynomial}

\item{extrapolation}{Either 1(linear) or 0(capped) used for extrapolating the RR after 150}
}
\description{
Get Simple Fractional Polynomial Relative Risk Function
}
