% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{vintegrate}
\alias{vintegrate}
\title{Integrate that takes as input a list funs of functions, a vector vlower of
"lower" values and a vector vupper of "upper" values, assumes that
length(vlower) = length(vupper) = length(funs), and returns a
vector of integrals of funs between the lower and upper values.}
\usage{
vintegrate(funs, vlower, vupper)
}
\description{
Integrate that takes as input a list funs of functions, a vector vlower of
"lower" values and a vector vupper of "upper" values, assumes that
length(vlower) = length(vupper) = length(funs), and returns a
vector of integrals of funs between the lower and upper values.
}
